refactor the way to pass the label component for divide logo animation
