+++ src/ui/buttons/buttons.jsx
@@ -8,7 +8,7 @@ import { CLASS, BUTTON_NUMBER, BUTTON_LAYOUT, BUTTON_FLOW } from '../../constant
 import { determineEligibleFunding, isWalletFundingEligible } from '../../funding';
 import { ValidationError } from '../../lib';
 
-import { createAddDivideLogoAnimation } from './buttonAnimations';
+import { createAddDivideLogoAnimation } from './button-animations';
 import { normalizeButtonProps, type ButtonPropsInputs, type OnShippingChange } from './props';
 import { Style } from './style';
 import { Button } from './button';
